% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2016.R
\name{convert2016}
\alias{convert2016}
\title{Convert NBI 2016 ASCII to Tibble}
\usage{
convert2016(filePath = NULL)
}
\arguments{
\item{filePath}{location of the NBI 2016 ASCII CSV file}
}
\value{
Tibble containing QC'ed NBI data.
}
\description{
A previously downloaded NBI ASCII CSV file for the 2016 version of 
\href{https://www.fhwa.dot.gov/bridge/nbi/2016hwybronlyonefile.zip}{Highway Bridges for all States}
is converted to a tibble with careful QC and renaming of the columns of data.

Columns in the returned tibble include:
\enumerate{
\item{stateCode} -- ISO 3166-2 alpha-2
\item{longitude} -- decimal degrees E
\item{latitude} -- decimal degrees N
}
}
\details{
The NBI 2016 dataset can be downloaded from the Federal Highway Administration
with the following command:
\preformatted{
curl https://www.fhwa.dot.gov/bridge/nbi/2016hwybronlyonefile.zip -O
}
This file should be left in '.zip' format.
}
\references{
\href{https://www.fhwa.dot.gov/bridge/nbi/ascii.cfm}{NBI ASCII data}

\href{https://www.fhwa.dot.gov/bridge/mtguide.pdf}{NBI Recording and Coding Guide}
}
\keyword{datagen}
